Update MainComponent.h
Included the base class Listener so that we can monitor and respond to changes in the sliders. I started to implement behavior for the sample frequency slider. It's probably not the right approach but it's just an example while I still try to understand what each slider will be changing.